Fixed config cache size 64bit issue.

+++ src/Framework/Storage/StorageConfig.cpp
@@ -5,5 +5,5 @@ void StorageConfig::Init()
 {
     chunkSize = configFile.GetIntValue("database.chunkSize", STORAGE_DEFAULT_CHUNKSIZE);
     logSegmentSize = configFile.GetIntValue("database.logSegmentSize", STORAGE_DEFAULT_LOGSEGMENTSIZE);
-    fileChunkCacheSize = configFile.GetIntValue("database.cacheSize", STORAGE_DEFAULT_FILECHUNK_CACHESIZE);
+    fileChunkCacheSize = (uint64_t) configFile.GetInt64Value("database.cacheSize", STORAGE_DEFAULT_FILECHUNK_CACHESIZE);
 }
